List favorite games and add/remove them

+++ Dysnomia.CoursFrontM1.GamesDb.Common/Dao/UserDao.cs
@@ -3,6 +3,6 @@
         public Guid Id { get; set; }
         public string Name { get; set; } = null!;
         public string HashedPassword { get; set; } = null!;
-        public List<int> Favorites { get; set; } = null!;
+        public List<ulong> Favorites { get; set; } = null!;
     }
 }
